Creating an Airplane(s)
Fixed the Advanced button to hide back on click.
Added a Menu bar with some default options, most of which don't work because are not implemented.
Only one wokrs which is in Tools->Create
Added GUI part for creating airplanes with a posibility of creating a bulk of airplanes
Added 2 helper methods in the Network for ID's.