Added check when airplanes have to depart + fixes to creating a problem
Airplanes check if it's their time to depart depending on the flight departure time and the global time variable.
Planes can no longer be created to airports which have problems.
Planes with flights inbound or outbound to an airport with a problem are cancelled (Status is set to LANDED).
Added check to RerouteAirplaneToNearestAirport for problems in the airport.